backend/requirements.txt
amqp==5.1.1
asgiref==3.7.2
bcrypt==4.0.1
billiard==4.1.0
celery==5.3.4
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
Django==4.2.7
django-cors-headers==4.2.0
django-nested-admin==4.0.2
djangorestframework==3.14.0
freezegun==1.2.2
idna==3.4
kombu==5.3.2
mysqlclient==2.2.0
Naked==0.1.32
paytmchecksum==1.7.0
Pillow==10.0.0
prompt-toolkit==3.0.39
psycopg2==2.9.9
pycrypto==2.6.1
python-dateutil==2.8.2
python-monkey-business==1.0.0
pytz==2023.3
PyYAML==6.0.1
redis==4.6.0
requests==2.31.0
shellescape==3.8.1
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
urllib3==2.1.0
vine==5.0.0
wcwidth==0.2.6
